add Table parser
Abstract away Ibope's functionality into the new TableParser class.
TableParser deals with general cases where polls are in tables. It tries
to locate a "Total" cell in the table, and get two columns: one that's
on the left of the "Total" cell (i.e. the candidate names) and another
that's under the "Total" cell (i.e. the candidate percentages).

The Datafolha parser will also use TableParser.