interface: add stop marker and postprocessing
Add two new features to the interface. The stop marker stops tells the
parser to stop parsing when it hits some text. This is useful with
parsers like Ibope 2012, where predicting whether we're done is hard
through already_have, but it's easy through some text -- indeed, Ibope
2012 PDFs stop reporting polls once the question about rejection gets
asked.

Postprocessing, on the other hand, is useful to handle things with the
data like deduplication of scenarios, and merging vereador polls that
are spread across multiple pages.