table: fix parsing bugs
Idiotically, TableParser was not handling get_pste() output properly.
This is what you get when you don't have unit tests.

Further, this addresses an issue with the threshold for detecting a
multiline TOTAL (which, if not addressed, stopped the entire Ibope 2012
parser from getting the first entry in any table). In the future, it
should probably be up for the IbopeParser itself to set this threshold.